backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
sqlite-python==0.0.1
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
requests==2.31.0

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Data Caching
python-cache==0.6.1

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# JSON
orjson==3.9.15

# Environment Variables
python-dotenv==1.0.0

# HTTPX for async requests
httpx==0.26.0

# CORS
starlette==0.35.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
